Dispense with a couple of TODOs
Removed several outdated TODOs and fix those that are easy to fix.
